Add code test for enum class as compile time index in std::array
Seems to work. User declares enum class and provides that as template
argument to the recorder. Recording can than be done using the provided
enum class as argument. Using the wrong enum class causes compiler
error, very good for preventing mistakes. Only problem is finding out
the number of values in the enum, and make sure the user do not set the
values explicitly in the enum class.

By creating a template helper function that at compile time checks for
out-of-range indexing in the array it should not cause any problems with
unique enum values, as long the as the maximum element is named "Count"
or similar. Not providing that element in the enum class is also a
compile time error.