[refactor] Have record method take arrays to record more data
Instead of only allowing a single value to be recorded at the time,
using initializer lists et al the package sent can be squeezed to hold
3x64-bit of data. So the user can than choose between value 2-value and
3-value data. The change also extends the Item type with additional
members in the union.

The unit type is ONLY set at the first value recorded from now on to
simplify the code. This forces the receiver to store and handle the data
type for each package, but that is a reasonable trade off, parsing the
package to extract the time or fetching the type from a local storage
container should be equally fast.