Switch storage class for socket and send buffer
Each recorder should have its own sending buffer so that the buffer can
be sent as a whole with the payload frame describing the meta data. That
removes the need for putting the recorder identifier into each data
package. Most importantly it will make it easier at the receiving
end. For each package only a single recorder data structure needs to be
fetched and updated. Rather than having to parse each item in the data
and for each of them fetch the corresponding recorder.

Making the socket static thread_local will minimize the amount of
sockets opened. It will now be limited to the amount of threads in the
application, which will be less than (or equal to) the amount of
recorders. (Having a single recorder per thread is a very special case).