Remove recorder_id from data item struct
The payload for a data item should not need to carry the recorder
id. Adding it to each item will just add a lot of extra data. So instead
the payload frame, i.e. the initial frame in a message should carry that
information.

This also removes the necessity to mask a single char for squeezing both
type and length into the same variable. (Perhaps some other bit masking
technique would have been better but now it is not necessary at all).